refactor: 관련 예외 클래스를 하나의 파일에 모아 관리하고 코드 구조를 단순화
