--- src/main/kotlin/com/jaeyeon/blackfriday/common/global/BlackFridayException.kt
@@ -1,8 +1,0 @@
-package com.jaeyeon.blackfriday.common.global
-
-import com.jaeyeon.blackfriday.common.exception.ErrorCode
-
-open class BlackFridayException(
-    val errorCode: ErrorCode,
-    message: String? = errorCode.message,
-) : RuntimeException(message)
